Fix regression in cc7e26b0
The previous patch created a resync loop by clobbering target directory when
the sync had completed; this was due to a leading slash not being present when
comparing the key to the cached key name in _is_deleted().

A better place to map "/" to "" is just before calling list_folder().